feat(prettier): Add Prettier formatter
Prettier automatically formats code, which ensures consistent code
format across the codebase. This is highly simplifies code reviews and
limits the amount of core format nits in reviews.

Unfortunatelly Prettier does not work by default with standardjs so to
get the best of both worlds this adds custom eslint config which uses
standardjs config and prettier config which ensures that standardjs
rules which conflict with prettier are removed. In addition
eslint-plugin-prettier adds Prettier rules for linting.

To prettify code run 'npm run prettier' or use prettier plugin in editor
of your choice.

Closes #26