feat(MoreInformation): introduce the MoreInformation compoment
affects: patternfly-react

The component consists of "More Information" link preceded with the fa-angle-right icon.
On click, inner components with additional text are rendered. The preceding icon is changed
to the fa-angle-down.
Subsequent mouse click collapses the content again.

The component is used either for texts exceeding size of tooltips or if the additional
information is expected to be copied and pasted.