feat(ExpandCollapse): introduce ExpandCollapse component
affects: patternfly-react

Implementation for the Expand Collapse Section is provided.

The component can be farther used for the "More Information" use-case,
where additional text/components are hidden unless the user clicks
on a link.

Such a per-request rendered content is used for texts exceeding scope of a tooltip or
when the user is expected to copy and paste part of the information.