fix(TextInput/Area): update type definitions
The "onChange" method receives two arguments: value and event.

According to TextArea.d.ts and TextInput.d.ts the value is
boolean while it is string.

This patch fixes this to the correct type.

In addition, "onChange" should be optional (TextArea.d.ts set
it to be mandatory by mistake).

Signed-off-by: Boaz Shuster <boaz.shuster.github@gmail.com>