fix(DataList): Add aria-label and aria-labelledby improvements (#1347)
* fix(DataList): Add aria-label and aria-labelledby improvements

The DataListToggle aria label now defaults to "Details." If the consumer provides an aria label, there is no
aria-labelledby. If the consumer uses the default aria label, aria-labelledby will be in the
following format: aria-labelledby=`${aria-labelledby of DataListItem} ${id}`. Users can also add an id prop to DataListContent and a corresponding aria-controls prop to DataListToggle.

Affects #1163

* refactor(DataList): Change buttons so they use ButtonVariant.plain rather than "plain."

Fixes #1163

* refactor(DataList): Convert classes back into consts with object destructuring

Fixes #1163