feat(Modal.tsx): Add targetDocument to Modal to allow showing the Modal in another window
Current implemention will not allow use of the Modal when the running
code is within an iframe. In order to allow the Modal to escape the
iframe, which is probably invisible, the iframe's parent window's
document must be used by the Modal itself. With the change, the Modal
will be able to create and manage the React portal within another frame.

Fixes #2551