Ensure PagerHeaderToolsGroup visibility mods are optional
In the official demos here:
https://www.patternfly.org/v4/documentation/react/demos/pagelayout

You can see from the horizontal nav example that only the default and lg
props are specified.

This is legal in JS and I assume is how it is expected to work. Set a
default and add exceptions for other sizes you'd like for other
properties.

In ts, there is a hard requirement on setting each of the properties
because it is typechecked unlike JS which is super tedious (especially
if you just want something to always be visible).

This brings the convenience of the JS examples to TS as well.