fix(Select): fix crash when opened using Enter with disabled options
Closes #5257

When a <Select> has a disabled options on top of the options list,
`this.refCollection[0]` will be an array in the shape of `[null, null]`.
This means that trying to call `this.refCollection[0][0].focus()`
will fail with TypeError in this case.

This patch fixes this issue, by finding the first non-null ref in the
collection and focusing on it, instead of assuming the first ref in the
collection is valid.