fix(Select): adjust Select component so that it also works with children changes while it's mounted
In Cockpit [1] we utilize this component in order to create a filesystem
browser field inside dialogs.

Initially the values of the SelectOptions are the subdirs of the root
('/') directory:

So:
<Select>
    <SelectOption value="bin" />
    <SelectOption value="boot" />
    etc.
</Select>

When the user selects one of the directories/SelectOptions we basically
navigate them inside the directory they selected, therefore we update all
the SelectOptions in order to reflect the selected directory content.

Of course it's expected that the current value of the typeahead select
input will still act as a filter for the new children.

This commit fixes exactly this issue.

In other words, now the filtering of the children is now done not only
in the onChange method, which is caused by user keyboard interaction,
but also when the component just updates.

One problematic part, was that the onFilter exposed property expect an
Input element event to parse, which of course does not exist when the
filtering was requested because of componentDidUpdate lifecycle method.

For now, onFilter will not be taking any effect.

[1] cockpit-project.org

Fixes #5374