DualListSelect: add onAvailableOptionsSelectInputChanged and onChosenOptionsSelectInputChanged
The use case can be as seen in the linked issue, that the cunsumer might
need to use the left search bar to query an API for available options.

This is understandable as the list of available options can get
really long in some use-cases and in that case the API calls should only
fetch the options that are relevant to the user input.

Fixes #5436