fix(Progress): make proper usage of accessibility properties
commit c2cb799 ensured that there is a warning when none of `title` and
`aria-label` properties are present for the Progress component.

However, when `aria-label` is used, the `aria-labbelledby` of the
Progress component in the DOM keeps pointing to a hidden description element.

This is an accessibility issue and is addressed by this commit, by
always ensuring that `aria-labbelledby` field will always point to an
existing progress label. If no such label exists aria-label can be
used.