fix(Checkbox): fix warning about ouiaSafe property
The checkbox was throwing the following warning:

Warning: React does not recognize the `ouiaSafe` prop on a DOM element.
If you intentionally want it to appear in the DOM as a custom attribute, spell it as lowercase `ouiasafe` instead.
If you accidentally passed it from a parent component, remove it from the DOM element.
...
    in input (created by Checkbox)
    in div (created by Checkbox)