TreeView: react to default expansion state change
While it was possible to render a TreeView with a subset of nodes
expanded, it was not possible to programmatically change the default
expansion of nodes (as in response to a search) as the underlying node
did not react to changes on the default expansion state. This patch adds
that functionality.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>