feat(react-console): accept random children in AccessConsole component
So far we were supporting children of three types,
SerialConsole, DesktopViewer and VncConsole.

This component is consumed by `cockpit machines` product and it's
used for exposing consoles of Virtual Machines.
A Virtual Machines however can have multiple consoles of one type; for example 2
different serial consoles, something which was not supported until this
commit.

For this reason we need to extend this component to accept also custom
children apart from a single from each from the predefined types.

Also start respecting the order of the children to display the options
in the AccessConsoles dropdown.