fix(Modal): remove aria-hidden when unmount
The component adds the aria-hidden attribute to its siblings when open
and removes it when closed. However, if the component is mounted as open
but then directly unmounted, its siblings will remain hidden from the
accessibility API forever.

Thus, the logic for removing these attributes is now also triggered when
the component is unmounted.