test(NumberInput): drop aria-describedby workaround
Test was setting the `aria-describedby=''` attribute to a NumericInput
inside an InputGroup, most probably to avoid the autogenerated value
(see https://github.com/patternfly/patternfly-react/issues/9667).

However, after cleaning up the InputGroup component,
`aria-describedby=''` is not a workaround anymore for unset the
attribute, breaking the snapshot associated to the test.