chore(docs): Add onClick to SkipToContent in demos (#11297)
* chore(docs): Add onClick to SkipToContent in demos

SkipToContent does not work in the context of the PatternFly page. Using document.getElementById sidesteps this issue and allows them to work as usual. I tried using refs, but because Page is an older class component, the refs were not compatible. We'd likely need to rewrite page as a function component. This is good enough for now, in my opinion. SkipToContentBasic works as advertised, so I am not touching that, and I also did not touch react-integration or unit tests.

* chore(SkipToContent): Update demo

Use refs in demo to match other demos.