chore(Tabs): Update separate content examples
Separate content examples did not make it clear enough that the hidden attribute on TabContent is used to set the initial state of Tabs with separate content. I updated the example so it is managed programmatically and updated the docs. I also added some tests to make sure it is working as advertised.